dom/base/test/test_bug562652.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=562652
-->
<head>
  <title>Test for Bug 562652</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=562652">Mozilla Bug 562652</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<div id="testtarget">_</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 562652 **/


var testCount = 0;
function createHTMLDoc() {
  var dtd = document.implementation.createDocumentType("HTML", "-//W3C//DTD HTML 4.01//EN", null);
  var d = document.implementation.createDocument(null, null, dtd);
  d.appendChild(d.createElement("html"));
  d.documentElement.appendChild(d.createElement("body"));
  d.body.setAttribute("id", "testtarget");
  return d;
}

function test(d) {
  var t = d.getElementById("testtarget");
  d.addEventListener("DOMNodeInserted", function(e) { ++testCount; });
  t.innerHTML = "_";
}

function runTests() {
  test(document);
  test(createHTMLDoc());
  is(testCount, 2, "DOMNodeInserted should have fired 2 times!");
  SimpleTest.finish();
}

addLoadEvent(runTests);

SimpleTest.waitForExplicitFinish();

</script>
</pre>
</body>
</html>
